third_party/WebKit/LayoutTests/shadow-dom/focus-navigation-slot-nested-fallback.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script src='resources/shadow-dom.js'></script>
<script src='resources/focus-utils.js'></script>
<div id='log'></div>
<html>
<body>
<p>Tests for moving focus by pressing tab key across shadow boundaries.<br>
To manually test, press tab key five times then shift+tab six times.<br>
It should traverse focusable elements in the increasing numerical order and then in the reverse order.</p>
<style>

#test-content > * {
    display: block;
    width: 400px;
    height: 150px;
    border: solid 1px black;
    margin: 10px;
    float: left;
}

pre {
    padding-top: 1rem;
    clear: both;
}

</style>
<input id='init'>
<div id='host'>
  <template data-mode='open'>
    <slot name="slot1">Slot shouldn't be focused.
      <slot name="slot2">Slot shouldn't be focused.
        <input id='non1'>This text shouldn't appear.</input>
      </slot>
      <slot name="slot3">Slot shouldn't be focused.
        <input id='second'>2. No host child is assigned to slot3.</input>
      </slot>
    </slot>

    <input id='third'>3. Inner Shadow Host.
      <template data-mode='open'>
        <slot name="slot4">Slot shouldn't be focused.
          <slot name="slot5">Slot shouldn't be focused.
          <input id='non2'>This text shouldn't appear.</input>
          </slot>
        </slot>
      </template>
      <input id='fourth' slot="slot4">4. Assigned to slot4.</input>
      <input id='non3' slot="slot5">
        This text shouldn't appear. slot5 node is the ignored fallback content.</input>
    </input>

    <input id='fifth'>5. Inner Shadow Host.
      <template data-mode='open'>
        <slot name="slot6">Slot shouldn't be focused.
          <input id='non4'>This tetx shouldn't appear.</input>
        </slot>
      </template>
      <slot name="slot7" slot="slot6">Slot shouldn't be focused. Assigned to slot6.
        <input id='non5'>This text shouldn't appear.</input>
      </slot>
    </input>
  </template>Slot shouldn't be focused.
  <input id='first' slot="slot2">1. Assigned to slot2.</input>
  <input id='sixth' slot="slot7">6. Assigned to slot7 which is assigned to slot6.</input>
</div>
</input>
<script>
'use strict';

test(() => {
  convertTemplatesToShadowRootsWithin(host);

  let elements = [
       'init',
       'host/first',
       'host/second',
       'host/third',
       'host/third/fourth',
       'host/fifth',
       'host/sixth'
  ];

  assert_focus_navigation_forward(elements);
  elements.reverse();
  assert_focus_navigation_backward(elements);
}, 'Focus should cover assigned elements of an assigned slot espacially there are fallback contents.');

/*
var oldActiveElement = null;
function log()
{
    setTimeout(function () {
        var newActiveElement = document.activeElement;

        var shadowRoot = newActiveElement.shadowRoot || newActiveElement.closedShadowRoot;
        if (shadowRoot) {
            var activeElementInShadow = shadowRoot.activeElement;
            if (activeElementInShadow)
                newActiveElement = activeElementInShadow;
        }

        if (oldActiveElement == newActiveElement || newActiveElement == document.body)
            return;
        oldActiveElement = newActiveElement;
        document.querySelector('pre').textContent += newActiveElement.firstChild.textContent.trim() + '\n';
    }, 0);
}

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

window.onload = function () {
    convertTemplatesToShadowRootsWithin(host);
    document.onkeydown = log;

    for (var element of document.querySelectorAll('#test-content *')) {
        if (element instanceof HTMLIFrameElement)
            element.contentDocument.onkeydown = log;
        else
            element.onfocus = log;
    }

    host.focus();
    document.onkeydown();

    if (window.eventSender)
        moveFocusForward(6);
}

function moveFocusForward(focusCount)
{
    setTimeout(function () {
        if (!focusCount)
            return moveFocusBackward(5);
        eventSender.keyDown('\t');
        setTimeout(function () {
            moveFocusForward(focusCount - 1);
        }, 1);
    }, 1);
}

function moveFocusBackward(focusCount)
{
    setTimeout(function () {
        if (!focusCount) {
            document.getElementById('test-content').style.display = 'none';
            testRunner.notifyDone();
            return;
        }
        eventSender.keyDown('\t', ['shiftKey']);
        setTimeout(function () {
            moveFocusBackward(focusCount - 1);
        }, 1);
    }, 1);
}
*/


</script>
</body>
</html>
